+++ src/Areas/Aks/Commands/AksJsonContext.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+using AzureMcp.Areas.Aks.Commands.Cluster;
+using AzureMcp.Areas.Aks.Models;
+
+namespace AzureMcp.Commands.Aks;
+
+[JsonSerializable(typeof(ClusterListCommand.ClusterListCommandResult))]
+[JsonSerializable(typeof(Cluster))]
+[JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
+internal sealed partial class AksJsonContext : JsonSerializerContext;