+++ src/Areas/Aks/Options/Cluster/ClusterListOptions.cs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.Aks.Options.Cluster;
+
+public class ClusterListOptions : BaseAksOptions;