+++ src/Program.cs
@@ -59,6 +59,7 @@ internal class Program
             new AzureMcp.Areas.Tools.ToolsSetup(),
 
             // Register Azure service areas
+            new AzureMcp.Areas.Aks.AksSetup(),
             new AzureMcp.Areas.AppConfig.AppConfigSetup(),
             new AzureMcp.Areas.Authorization.AuthorizationSetup(),
             new AzureMcp.Areas.AzureIsv.AzureIsvSetup(),